+++ src/transport/Transport.java
@@ -1,0 +1,79 @@
+package transport;
+
+public class Transport {
+    private String brand;
+    private String model;
+    private int productionYear;
+    private String productionCountry;
+    private String color;
+    private double maxSpeed;
+
+    public Transport(String brand, String model, int productionYear, String productionCountry, String color, double maxSpeed) {
+        this.brand = brand;
+        this.model = model;
+        this.productionYear = productionYear;
+        this.productionCountry = productionCountry;
+        setColor(color);
+        setMaxSpeed(maxSpeed);
+        if (brand == null || brand.isEmpty()) {
+            this.brand = "default";
+        } else {
+            this.brand = brand;
+        }
+        if (model == null || model.isEmpty()) {
+            this.model = "default";
+        } else {
+            this.model = model;
+        }
+        if (productionCountry == null || productionCountry.isEmpty()) {
+            this.productionCountry = "default";
+        } else {
+            this.productionCountry = productionCountry;
+        }
+        if (productionYear <= 0) {
+            this.productionYear = 2000;
+        } else {
+            this.productionYear = productionYear;
+        }
+    }
+
+    public String getBrand() {
+        return brand;
+    }
+
+    public String getModel() {
+        return model;
+    }
+
+    public int getProductionYear() {
+        return productionYear;
+    }
+
+    public String getProductionCountry() {
+        return productionCountry;
+    }
+
+    public String getColor() {
+        return color;
+    }
+
+    public void setColor(String color) {
+        if (color == null || color.isEmpty()) {
+            this.color = "белый";
+        } else {
+            this.color = color;
+        }
+    }
+
+    public double getMaxSpeed() {
+        return maxSpeed;
+    }
+
+    public void setMaxSpeed(double maxSpeed) {
+        if (maxSpeed == 0) {
+            System.out.println(" не указана.");
+        } else {
+            this.maxSpeed = maxSpeed;
+        }
+    }
+}
